requirements.txt
amqp==2.5.0
Babel==2.7.0
billiard==3.6.0.0
boto3==1.9.196
botocore==1.12.196
celery==4.3.0
certifi==2019.6.16
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.3
django-allauth==0.39.1
django-countries==5.3.3
django-crispy-forms==1.7.2
django-phonenumber-field==3.0.1
django-storages==1.7.1
docutils==0.14
gunicorn==19.9.0
idna==2.8
jmespath==0.9.4
kombu==4.6.3
oauthlib==3.0.2
phonenumbers==8.10.15
Pillow==6.1.0
psycopg2-binary==2.8.3
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2019.1
redis==3.2.1
requests==2.22.0
requests-oauthlib==1.2.0
s3transfer==0.2.1
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
vine==1.3.0
